+++ src/Middlewares/AddJwtFieldsMiddleware.php
@@ -4,15 +4,10 @@ namespace Carsguide\Auth\Middlewares;
 
 use Auth0\SDK\Exception\CoreException;
 use Auth0\SDK\Exception\InvalidTokenException;
-use Auth0\SDK\Helpers\JWKFetcher;
-use Auth0\SDK\Helpers\Tokens\AsymmetricVerifier;
-use Auth0\SDK\Helpers\Tokens\IdTokenVerifier;
-use Carsguide\Auth\Handlers\CacheHandler;
 use Closure;
-use Exception;
 use Illuminate\Http\JsonResponse;
 
-class AddJwtFieldsMiddleware
+class AddJwtFieldsMiddleware extends BaseAuthMiddleware
 {
     /**
      * Run the request filter.
@@ -42,31 +37,6 @@ class AddJwtFieldsMiddleware
     }
 
     /**
-     * verify and decode the token
-     *
-     * @param string $token
-     * @return bool|void
-     */
-    public function verifyAndDecodeToken($token)
-    {
-        $auth0Domain = $this->getAuth0Domain();
-
-        if (!$auth0Domain || filter_var($auth0Domain, FILTER_VALIDATE_URL) === false) {
-            return false;
-        }
-
-        $jwksUri = $auth0Domain . '.well-known/jwks.json';
-
-        $jwksFetcher = new JWKFetcher(new CacheHandler(), [ 'base_uri' => $jwksUri ]);
-        $jwks        = $jwksFetcher->getKeys();
-        $sigVerifier = new AsymmetricVerifier($jwks);
-
-        $idTokenVerifier = new IdTokenVerifier($auth0Domain, env('AUTH0_AUDIENCE', false), $sigVerifier);
-
-        $this->decodedToken = $idTokenVerifier->verify($token);
-    }
-
-    /**
      * Explode fields
      *
      * @param $fields
@@ -88,18 +58,6 @@ class AddJwtFieldsMiddleware
             $value = $this->decodedToken[$field] ?? null;
             $this->request->merge([$field => $value]);
         }
-    }
-
-    /**
-     * If multiple Auth0 domain is set, return the first one
-     *
-     * @return string
-     */
-    protected function getAuth0Domain(): string
-    {
-        $auth0Domains = explode(',', env('AUTH0_DOMAIN', false));
-
-        return $auth0Domains[0] ?? '';
     }
 
     /**
