+++ src/Middlewares/AddJwtNamespaceFieldsMiddleware.php
@@ -4,10 +4,6 @@ namespace Carsguide\Auth\Middlewares;
 
 use Auth0\SDK\Exception\CoreException;
 use Auth0\SDK\Exception\InvalidTokenException;
-use Auth0\SDK\Helpers\JWKFetcher;
-use Auth0\SDK\Helpers\Tokens\AsymmetricVerifier;
-use Auth0\SDK\Helpers\Tokens\IdTokenVerifier;
-use Carsguide\Auth\Handlers\CacheHandler;
 use Closure;
 use Illuminate\Http\JsonResponse;
 
@@ -38,31 +34,6 @@ class AddJwtNamespaceFieldsMiddleware
         $this->addFieldsToRequest();
 
         return $next($this->request);
-    }
-
-    /**
-     * verify and decode the token
-     *
-     * @param string $token
-     * @return bool|void
-     */
-    public function verifyAndDecodeToken($token): bool
-    {
-        $auth0Domain = $this->getAuth0Domain();
-
-        if (!$auth0Domain || filter_var($auth0Domain, FILTER_VALIDATE_URL) === false) {
-            return false;
-        }
-
-        $jwksUri = $auth0Domain . '.well-known/jwks.json';
-
-        $jwksFetcher = new JWKFetcher(new CacheHandler(), [ 'base_uri' => $jwksUri ]);
-        $jwks        = $jwksFetcher->getKeys();
-        $sigVerifier = new AsymmetricVerifier($jwks);
-
-        $idTokenVerifier = new IdTokenVerifier($auth0Domain, env('AUTH0_AUDIENCE', false), $sigVerifier);
-
-        $this->decodedToken = $idTokenVerifier->verify($token);
     }
 
     /**
@@ -102,18 +73,6 @@ class AddJwtNamespaceFieldsMiddleware
     protected function getJwtNamespace()
     {
         return env('AUTH0_JWT_NAMESPACE', 'https://platform.autotrader.com.au/');
-    }
-
-    /**
-     * If multiple Auth0 domain is set, return the first one
-     *
-     * @return string
-     */
-    protected function getAuth0Domain(): string
-    {
-        $auth0Domains = explode(',', env('AUTH0_DOMAIN', false));
-
-        return $auth0Domains[0] ?? '';
     }
 
     /**
