+++ src/Middlewares/Auth0Middleware.php
@@ -4,16 +4,12 @@ namespace Carsguide\Auth\Middlewares;
 
 use Auth0\SDK\Exception\CoreException;
 use Auth0\SDK\Exception\InvalidTokenException;
-use Auth0\SDK\Helpers\JWKFetcher;
-use Auth0\SDK\Helpers\Tokens\AsymmetricVerifier;
-use Auth0\SDK\Helpers\Tokens\IdTokenVerifier;
-use Carsguide\Auth\Handlers\CacheHandler;
 use Closure;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Log;
 
-class Auth0Middleware
+class Auth0Middleware extends BaseAuthMiddleware
 {
     /**
      * Run the request filter.
@@ -66,30 +62,7 @@ class Auth0Middleware
         return $next($request);
     }
 
-    /**
-     * verify and decode the token
-     *
-     * @param string $token
-     * @return bool|void
-     */
-    public function verifyAndDecodeToken($token)
-    {
-        $auth0Domain = $this->getAuth0Domain();
 
-        if (!$auth0Domain || filter_var($auth0Domain, FILTER_VALIDATE_URL) === false) {
-            return false;
-        }
-
-        $jwksUri = $auth0Domain . '.well-known/jwks.json';
-
-        $jwksFetcher = new JWKFetcher(new CacheHandler(), [ 'base_uri' => $jwksUri ]);
-        $jwks        = $jwksFetcher->getKeys();
-        $sigVerifier = new AsymmetricVerifier($jwks);
-
-        $idTokenVerifier = new IdTokenVerifier($auth0Domain, env('AUTH0_AUDIENCE', false), $sigVerifier);
-
-        $this->decodedToken = $idTokenVerifier->verify($token);
-    }
 
     /**
      * verify the user has access to scope and scopes are defined in JWT
@@ -119,19 +92,6 @@ class Auth0Middleware
     public function getScopesFromToken()
     {
         return explode(' ', $this->decodedToken['scope']);
-    }
-
-
-    /**
-     * If multiple Auth0 domain is set, return the first one
-     *
-     * @return string
-     */
-    protected function getAuth0Domain(): string
-    {
-        $auth0Domains = explode(',', env('AUTH0_DOMAIN', false));
-
-        return $auth0Domains[0] ?? '';
     }
 
     /**
